Add RootFinder type class approach to root finding.
