+++ core/src/main/scala/spire/math/Polynomial.scala
@@ -156,170 +156,68 @@ object Polynomial extends PolynomialInstances {
     loop(Polynomial.zero[C], Nil, points.toList)
   }
 
-  private implicit object JBigDecimalOrder extends Signed[JBigDecimal] with Order[JBigDecimal] {
-    def signum(a: JBigDecimal): Int = a.signum
-    def abs(a: JBigDecimal): JBigDecimal = a.abs
-    def compare(x: JBigDecimal, y: JBigDecimal): Int = x compareTo y
-  }
+  /**
+   * A trait that can be used to retreive the (possibly approximated) real
+   * roots of the polynomial `poly`.
+   */
+  trait Roots[A] { self =>
 
-  def QIR(poly: Polynomial[Rational], lowerBound: Rational, upperBound: Rational, digits: Int): Interval[BigDecimal] = {
+    /** The polynomial the roots belong to. */
+    def poly: Polynomial[A]
 
-    // TODO: See if Algebraic is faster.
-    def evalExact(x: JBigDecimal): JBigDecimal =
-      poly(Rational(new BigDecimal(x, MathContext.UNLIMITED)))
-        .toBigDecimal(digits, RoundingMode.UP)
-        .bigDecimal
+    /** Returns the number of real roots of `poly`. */
+    def count: Int
 
-    val lb = lowerBound.toBigDecimal(digits, RoundingMode.CEILING).bigDecimal
-    val ub = upperBound.toBigDecimal(digits, RoundingMode.FLOOR).bigDecimal
-
-    val qlb = Rational(new BigDecimal(lb, MathContext.UNLIMITED))
-    val qub = Rational(new BigDecimal(ub, MathContext.UNLIMITED))
-
-    // Returns true if there is a root in the open sub-interval (l, r).
-    def hasRoot(l: Rational, r: Rational): Boolean =
-      if (l != r) {
-        // Ue Descartes' rule of signs to see if the root in the open interval
-        // is actually in the sub interval (l, r).
-        val poly0 = poly
-          .shift(l)
-          .removeZeroRoots
-          .reciprocal
-          .shift((r - l).reciprocal)
-          .removeZeroRoots
-        poly0.signVariations % 2 == 1
-      } else {
-        false
-      }
-
-    if (hasRoot(lowerBound, qlb)) {
-      Interval.point(BigDecimal(lb))
-    } else if (hasRoot(qub, upperBound)) {
-      Interval.point(BigDecimal(ub))
-    } else {
-      QIR(lb, ub, digits)(evalExact).mapBounds(BigDecimal(_))
-    }
+    /**
+     * Returns the `i`-th real root of `poly`, or throws an
+     * `IndexOutOfBoundsException` if there is no `i`-th real root.
+     */
+    def get(i: Int): A
   }
 
   /**
-   * An implementation of "Quadratic Interval Refinement for Real Roots" by
-   * John Abbot for `BigDecimal`.
+   * A type class that can find roots of a polynomial.
    */
-  def QIR(lowerBound: JBigDecimal, upperBound: JBigDecimal, digits: Int)(evalExact: JBigDecimal => JBigDecimal): Interval[JBigDecimal] = {
-    val eps = JBigDecimal.valueOf(1, digits)
+  trait RootFinder[A] {
 
-    @tailrec
-    def loop(
-      lx: JBigDecimal,
-      ly: JBigDecimal,
-      rx: JBigDecimal,
-      ry: JBigDecimal,
-      n: Int
-    ): Interval[JBigDecimal] = {
-      val dy = ly.subtract(ry)
-      val s = ly.divide(dy, n, RoundingMode.HALF_UP)
-      val dx = rx.subtract(lx)
-      if (dx.compareTo(eps) <= 0) {
-        Interval.open(lx, rx)
-      } else {
-        val delta = dx.multiply(s.ulp)
-        val k = s.unscaledValue
-        val x1 = lx.add(delta.multiply(new JBigDecimal(k))).setScale(digits, RoundingMode.HALF_UP)
-        val y1 = evalExact(x1)
-        val s1 = y1.sign
-        if (s1 == ly.sign) {
-          val x2 = x1.add(delta).setScale(digits, RoundingMode.CEILING)
-          val y2 = evalExact(x2)
-          val s2 = y2.sign
-          if (s2 == s1) loop0(lx, ly, rx, ry)
-          else if (s2 == ry.sign) loop(x1, y1, x2, y2, 2 * n)
-          else Interval.point(x2)
-        } else if (s1 == ry.sign) {
-          val x0 = x1.subtract(delta).setScale(digits, RoundingMode.FLOOR)
-          val y0 = evalExact(x0)
-          val s0 = y0.sign
-          if (s0 == s1) loop0(lx, ly, rx, ry)
-          else if (s0 == ly.sign) loop(x0, y0, x1, y1, 2 * n)
-          else Interval.point(x0)
-        } else {
-          Interval.point(x1)
-        }
+    /**
+     * Returns the roots of the polynomial `poly`.
+     */
+    def findRoots(poly: Polynomial[A]): Roots[A]
+  }
+
+  object RootFinder {
+    implicit def BigDecimalScaleRootFinder(scale: Int): RootFinder[BigDecimal] =
+      new RootFinder[BigDecimal] {
+        def findRoots(poly: Polynomial[BigDecimal]): Roots[BigDecimal] =
+          new BigDecimalRoots(poly, scale)
+      }
+  }
+
+  private class BigDecimalRoots(
+    val poly: Polynomial[BigDecimal],
+    scale: Int
+  ) extends Roots[BigDecimal] {
+    private val qPoly: Polynomial[Rational] = poly.map(Rational(_))
+    private val isolated: Vector[Interval[Rational]] = isolateRoots(qPoly)
+
+    def count: Int = isolated.size
+
+    def get(i: Int): BigDecimal = if (i < 0 || i >= count) {
+      throw new IndexOutOfBoundsException(i.toString)
+    } else {
+      isolated(i) match {
+        case Point(value) =>
+          value.toBigDecimal(scale, RoundingMode.HALF_EVEN)
+        case Bounded(lb, ub, _) =>
+          new BigDecimal(
+            BigDecimalRootRefinement.QIR(qPoly, lb, ub, scale).approximation,
+            MathContext.UNLIMITED
+          )
+        case _ =>
+          throw new RuntimeException("invalid isolated root interval")
       }
     }
-
-    def bisect(
-      x0: JBigDecimal,
-      y0: JBigDecimal,
-      x1: JBigDecimal,
-      y1: JBigDecimal,
-      x2: JBigDecimal,
-      y2: JBigDecimal
-    ): Interval[JBigDecimal] = {
-      if (y0.signum == 0) Interval.point(x0)
-      else if (y1.signum == 0) Interval.point(x1)
-      else if (y2.signum == 0) Interval.point(x2)
-      else if (y0.sign != y1.sign) loop(x0, y0, x1, y1, 1)
-      else loop(x1, y1, x2, y2, 1)
-    }
-
-    def loop0(
-      x0: JBigDecimal,
-      y0: JBigDecimal,
-      x5: JBigDecimal,
-      y5: JBigDecimal
-    ): Interval[JBigDecimal] = {
-      val dy = y0.subtract(y5)
-      val k = y0
-        .divide(dy, 1, RoundingMode.HALF_UP)
-        .unscaledValue
-        .intValue
-      val eps = x5.subtract(x0).divide(new JBigDecimal(5))
-      def eval(k: Int): (JBigDecimal, JBigDecimal) = {
-        val x = new JBigDecimal(k).multiply(eps).add(x0)
-        val y = evalExact(x)
-        (x, y)
-      }
-
-      if (k < 5) {
-        val (x2, y2) = eval(2)
-        if (y2.sign != y0.sign) {
-          val (x1, y1) = eval(1)
-          bisect(x0, y0, x1, y1, x2, y2)
-        } else {
-          val (x3, y3) = eval(3)
-          if (y3.sign == y5.sign) {
-            loop(x2, y2, x3, y3, 1)
-          } else {
-            val (x4, y4) = eval(4)
-            bisect(x3, y3, x4, y4, x5, y5)
-          }
-        }
-      } else {
-        val (x3, y3) = eval(3)
-        if (y3.sign != y5.sign) {
-          val (x4, y4) = eval(4)
-          bisect(x3, y3, x4, y4, x5, y5)
-        } else {
-          val (x2, y2) = eval(2)
-          if (y2.sign == y0.sign) {
-            loop(x2, y2, x3, y3, 1)
-          } else {
-            val (x1, y1) = eval(1)
-            bisect(x0, y0, x1, y1, x2, y2)
-          }
-        }
-      }
-    }
-
-    val lx = lowerBound
-    val ly = evalExact(lx)
-    val rx = upperBound
-    val ry = evalExact(rx)
-    val refined = loop0(lx, ly, rx, ry)
-    Interval.fromBounds(
-      refined.lowerBound.map(_.setScale(digits, RoundingMode.FLOOR)),
-      refined.upperBound.map(_.setScale(digits, RoundingMode.CEILING))
-    )
   }
 
   /**
@@ -482,6 +380,9 @@ trait Polynomial[@spec(Double) C] { lhs =>
     bldr.result()
   }
 
+  def roots(implicit finder: Polynomial.RootFinder[C]): Polynomial.Roots[C] =
+    finder.findRoots(this)
+
   /** Returns the coefficient of the n-th degree term. */
   def nth(n: Int)(implicit ring: Semiring[C]): C
 
@@ -558,6 +459,9 @@ trait Polynomial[@spec(Double) C] { lhs =>
     val Term(_, k) = minTerm
     mapTerms { case Term(c, n) => Term(c, n - k) }
   }
+
+  def map[D: Semiring: Eq: ClassTag](f: C => D)(implicit ring: Semiring[C], eq: Eq[C]): Polynomial[D] =
+    mapTerms { case Term(c, n) => Term(f(c), n) }
 
   def mapTerms[D: Semiring: Eq: ClassTag](f: Term[C] => Term[D])(implicit ring: Semiring[C], eq: Eq[C]): Polynomial[D] =
     Polynomial(terms map f)
